backend/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pymysql
cryptography
langchain
langchain-community
langchain-core
huggingface_hub
langchain-openai
langchain-anthropic
langchain_huggingface
langchain_groq
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator